Making cache_location mandatory
Minimize the chance of irrelative apps sharing tokens by removing the default centralized location.

Alternatively, we could keep the default behavior but use a per-app location such as os.path.join(sys.argv[0] + ".cache").
But we choose to NOT do this flavor this time. It's always easier to add surface area later than to take it away.